pyproject.toml
[tool.poetry]
name = "mycroft-roon-skill"
version = "0.0.1"
description = "mycroft/ovos roon skill plugin"
authors = ["Ramblurr <unnamedrambler@gmail.com>"]
license = "AGPL-3.0-or-later"
keywords = ["ovos", "skill", "plugin"]
packages = [
    { include = "roon_skill" },
    { include = "roon_proxy" },
    { include = "rpc" }
]

[tool.poetry.dependencies]
python = "^3.11"
roonapi = "0.1.4"
fuzzywuzzy = "0.18.0"
asyncio = "3.4.3"
msgspec = "0.18.1"
pyzmq = "25.1.1"
ovos-utils = "0.0.35"
ovos-workshop = "0.0.12"
adapt-parser = "1.0.0"

[tool.poetry.dev-dependencies]
flake8-docstrings = ">=1.3.0"
flake8 = ">=3.6.0"
flake8-quotes = ">=3.2.0"
pyright = ">=1.1.323"
pydocstyle = ">=3.0.0"
pylint = "==2.4.3"
pytest = "==5.2.1"
black = ">=22.12"

[tool.black]
line-length = 88

[tool.poetry.plugins."ovos.plugin.skill"]
"mycroft-roon-skill.ramblurr" = "mycroft-roon-skill:RoonSkill"

[tool.poetry.scripts]
roon-skill = "mycroft_roon_skill:RoonSkill"
roon-proxy-server = "roon_proxy.roon_proxy_server:main"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
